explicit cast to int

+++ sourceCode/scripting/scriptObject.cpp
@@ -1606,7 +1606,7 @@ int CScriptObject::_callSystemScriptFunction(int callType,const CInterfaceStack*
     // ---------------------------------
     if (_scriptType==sim_scripttype_mainscript)
     { // corresponding calls for plugins:
-        int data[4]={0,App::currentWorld->simulation->getSimulationTime_us()/1000,0,0};
+        int data[4]={0,int(App::currentWorld->simulation->getSimulationTime_us()/1000),0,0};
         if (callType==sim_syscb_init)
             CPluginContainer::sendEventCallbackMessageToAllPlugins(sim_message_eventcallback_simulationinit,data,nullptr,nullptr);
         if (callType==sim_syscb_actuation)
@@ -1636,7 +1636,7 @@ int CScriptObject::_callSystemScriptFunction(int callType,const CInterfaceStack*
     }
     if (_scriptType==sim_scripttype_mainscript)
     { // corresponding calls for plugins:
-        int data[4]={1,App::currentWorld->simulation->getSimulationTime_us()/1000,0,0};
+        int data[4]={1,int(App::currentWorld->simulation->getSimulationTime_us()/1000),0,0};
         if (callType==sim_syscb_init)
             CPluginContainer::sendEventCallbackMessageToAllPlugins(sim_message_eventcallback_simulationinit,data,nullptr,nullptr);
         if (callType==sim_syscb_actuation)
